+++ x-pack/solutions/observability/plugins/observability/public/pages/alert_details/components/related_dashboards/dashboard_tile.tsx
@@ -23,7 +23,7 @@ export interface ActionButtonProps {
   isDisabled: boolean;
 }
 
-export function Dashboard({
+export function DashboardTile({
   dashboard,
   actionButtonProps,
 }: {
